Fix locale association
When using multiple locales with e.g. a_locale.json files *and* folders with php files a_locale/errors.php then the translations were mixed in the wrong files because of a wrong "lastLocale". The locale will now be determined by the given file path. That results to the correct association.